+++ test/fixture/src/utils/icons.js
@@ -1,0 +1,244 @@
+export const ADJUNTAR = 'attach_file';
+export const ADJUNTOS = 'icon-adjuntos';
+export const ALERTA = 'icon-alerta';
+export const ALERTA1 = 'icon-alerta1';
+export const ANADIR = 'icon-anadir';
+export const ANADIR_BOX = 'icon-anadir-box';
+export const ANADIR_VARIOS = 'icon-anadir-varios';
+export const ANULAR = 'icon-anular';
+export const BASE_DATOS = 'icon-base-datos';
+export const BORRAR = 'icon-borrar';
+export const BORRAR_TODO = 'icon-borrar-todo';
+export const BUSCAR = 'icon-buscar';
+export const CALENDARIO = 'icon-calendario';
+export const CANCEL = 'icon-cancel';
+export const CANDADO = 'icon-candado';
+export const CARGADOR_EXPEDIENTE = 'icon-cargador-expediente';
+export const CARGADOR_EXPEDIENTES = 'icon-cargador-expedientes';
+export const CARGOS = 'icon-cargos';
+export const CARPETA = 'icon-carpeta';
+export const CERRAR = 'icon-cerrar';
+export const CERRAR_SESION = 'icon-cerrar-sesion';
+export const CHECK = 'icon-check';
+export const CODIGO_BARRAS = 'icon-codigo-barras';
+export const CODIGO_QR = 'icon-codigo-qr';
+export const COLECTIVO = 'icon-colectivo';
+export const COLECTIVOS = 'icon-colectivos';
+export const COLEGIADO = 'icon-colegiado';
+export const COMPRIMIR = 'icon-comprimir';
+export const CONSULTA = 'icon-consulta';
+export const CONVERTIR = 'icon-convertir';
+export const DATOS_PROCEDIMIENTO = 'icon-datos-procedimiento';
+export const DATOS_TRAMITE = 'icon-datos-tramite';
+export const DELITOS = 'icon-delitos';
+export const DESACUMULAR = 'icon-desacumular';
+export const DETALLE = 'icon-detalle';
+export const DETENIDO = 'icon-detenido';
+export const DEVOLUCIONES = 'icon-devoluciones';
+export const DOC_PRINCIPAL = 'icon-doc-principal';
+export const DOCUMENTOS = 'icon-documentos';
+export const DOCUMENTO_SUBIDA = 'icon-documento-subida';
+export const DOMICILIO = 'icon-domicilio';
+export const EDITAR = 'icon-editar';
+export const ENTIDAD = 'icon-entidad';
+export const ENVIOS = 'icon-envios';
+export const ERROR_CIRCULO = 'icon-error-circulo';
+export const ESCANEAR = 'icon-escanear';
+export const ESCRITO_TRAMITE = 'icon-escrito-tramite';
+export const EXCEL = 'icon-excel';
+export const FIGURA = 'icon-figura';
+export const FILTRAR = 'icon-filtrar';
+export const FIRMANTES = 'icon-pen-firmantes';
+export const FOLDER_OPEN = 'icon-folder-open';
+export const FUNCIONALIDADES = 'icon-funcionalidades';
+export const GENERAR_LISTADO = 'icon-generar-listado';
+export const GENERAR_PDF = 'icon-generar-pdf';
+export const HISTORICO = 'icon-historico';
+export const IMG = 'icon-img';
+export const IMPRIMIR = 'icon-Imprimir';
+export const INFO = 'icon-info';
+export const INFO_CIRCULO = 'icon-info-circulo';
+export const INVERTIR_ABAJO = 'icon-invertir-abajo';
+export const INVERTIR_ARRIBA = 'icon-invertir-arriba';
+export const ITINERACION = 'icon-itineracion';
+export const JUEZ = 'icon-juez';
+export const JUSTICIA = 'icon-justicia';
+export const LEXNET = 'icon-lexnet';
+export const LIMPIAR = 'icon-limpiar';
+export const MACROCAUSAS = 'icon-macrocausas';
+export const MAGISTRADO = 'icon-magistrado';
+export const MAPA_ASUNTO = 'icon-mapa-asunto';
+export const MENU = 'menu';
+export const MINUTA = 'icon-minuta';
+export const MINUTA_BOX = 'icon-minuta-box';
+export const MINUTA_EDIT = 'icon-minuta-edit';
+export const MOVER_ABAJO = 'icon-mover-abajo';
+export const MOVER_ARRIBA = 'icon-mover-arriba';
+export const MOVER_DERECHA = 'icon-mover-derecha';
+export const MOVER_IZQUIERDA = 'icon-mover-izquierda';
+export const NOTIFICACION = 'icon-notificacion';
+export const NOTIFICAR = 'icon-notificar';
+export const OBJETOS = 'icon-objetos';
+export const OCULTAR = 'icon-ocultar';
+export const ODT = 'icon-odt';
+export const ORIGEN = 'icon-origen';
+export const PENA_MEDIDA = 'icon-pena-medida';
+export const PEN_CUMPLIMENTAR = 'icon-pen-cumplimentar';
+export const PEN_CUMPLIMENTAR_SEALAMIENTO = 'icon-pen-cumplimentar-sealamiento';
+export const PEN_ENVIO_FIRMA = 'icon-pen-envio-firma';
+export const PEN_ENVIO_NOTIFICACION = 'icon-pen-envio-notificacion';
+export const PEN_ENVIO_SEALAMIENTO = 'icon-pen-envio-sealamiento';
+export const PEN_FIRMA_LOCAL = 'icon-pen-firma-local';
+export const PEN_MENSAJERIA = 'icon-pen-mensajeria';
+export const PEN_RESOLUCION = 'icon-pen-resolucion';
+export const PEN_TRAMITACION = 'icon-pen-tramitacion';
+export const PERSONA = 'icon-persona';
+export const PIN = 'icon-pin';
+export const POWERPOINT = 'icon-powerpoint';
+export const PREGUNTA = 'icon-pregunta';
+export const PREGUNTA_CIRCULO = 'icon-pregunta-circulo';
+export const PRESO = 'icon-preso';
+export const PROFESIONALES = 'icon-profesionales';
+export const RECEPCION = 'icon-recepcion';
+export const RECUPERAR = 'icon-recuperar';
+export const REFRESCAR = 'icon-refrescar';
+export const RELACIONES = 'icon-relaciones';
+export const SECRETARIO = 'icon-secretario';
+export const SELEC_DOC = 'icon-selec-doc';
+export const SIGUIENTE = 'icon-siguiente';
+export const SIN_FILTRO = 'icon-sin-filtro';
+export const TIPO_DOCUMENTAL = 'icon-tipo-documental';
+export const TRAMITE_OTRO_PROC = 'icon-tramite-otro-proc';
+export const TRAMITES = 'icon-tramites';
+export const UNSELECT_DOC = 'icon-unselect-doc';
+export const URGENTE = 'icon-urgente';
+export const VER = 'icon-ver';
+export const VIOLENCIA_GENERO = 'icon-violencia-genero';
+export const VISIBILITY = 'visibility';
+export const VOLVER = 'icon-volver';
+export const WORD = 'icon-word';
+export const ZIP = 'icon-zip';
+export const ZOOM_IN = 'zoom_in';
+
+export const ICONS = {
+  ADJUNTAR,
+  ADJUNTOS,
+  ALERTA,
+  ALERTA1,
+  ANADIR,
+  ANADIR_BOX,
+  ANADIR_VARIOS,
+  ANULAR,
+  BASE_DATOS,
+  BORRAR,
+  BORRAR_TODO,
+  BUSCAR,
+  CALENDARIO,
+  CANCEL,
+  CANDADO,
+  CARGADOR_EXPEDIENTE,
+  CARGADOR_EXPEDIENTES,
+  CARGOS,
+  CARPETA,
+  CERRAR,
+  CERRAR_SESION,
+  CHECK,
+  CODIGO_BARRAS,
+  CODIGO_QR,
+  COLECTIVO,
+  COLECTIVOS,
+  COLEGIADO,
+  COMPRIMIR,
+  CONSULTA,
+  CONVERTIR,
+  DATOS_PROCEDIMIENTO,
+  DATOS_TRAMITE,
+  DELITOS,
+  DESACUMULAR,
+  DETALLE,
+  DETENIDO,
+  DEVOLUCIONES,
+  DOC_PRINCIPAL,
+  DOCUMENTOS,
+  DOCUMENTO_SUBIDA,
+  DOMICILIO,
+  EDITAR,
+  ENTIDAD,
+  ENVIOS,
+  ERROR_CIRCULO,
+  ESCANEAR,
+  ESCRITO_TRAMITE,
+  FIGURA,
+  FILTRAR,
+  FIRMANTES,
+  FOLDER_OPEN,
+  FUNCIONALIDADES,
+  GENERAR_LISTADO,
+  GENERAR_PDF,
+  HISTORICO,
+  IMG,
+  IMPRIMIR,
+  INFO,
+  INFO_CIRCULO,
+  INVERTIR_ABAJO,
+  INVERTIR_ARRIBA,
+  ITINERACION,
+  JUEZ,
+  JUSTICIA,
+  LEXNET,
+  LIMPIAR,
+  MACROCAUSAS,
+  MAGISTRADO,
+  MAPA_ASUNTO,
+  MENU,
+  MINUTA,
+  MINUTA_BOX,
+  MINUTA_EDIT,
+  MOVER_ABAJO,
+  MOVER_ARRIBA,
+  MOVER_DERECHA,
+  MOVER_IZQUIERDA,
+  NOTIFICACION,
+  NOTIFICAR,
+  OBJETOS,
+  OCULTAR,
+  ODT,
+  ORIGEN,
+  PENA_MEDIDA,
+  PEN_CUMPLIMENTAR,
+  PEN_CUMPLIMENTAR_SEALAMIENTO,
+  PEN_ENVIO_FIRMA,
+  PEN_ENVIO_NOTIFICACION,
+  PEN_ENVIO_SEALAMIENTO,
+  PEN_FIRMA_LOCAL,
+  PEN_MENSAJERIA,
+  PEN_RESOLUCION,
+  PEN_TRAMITACION,
+  PERSONA,
+  PIN,
+  POWERPOINT,
+  PREGUNTA,
+  PREGUNTA_CIRCULO,
+  PRESO,
+  PROFESIONALES,
+  RECEPCION,
+  RECUPERAR,
+  REFRESCAR,
+  RELACIONES,
+  SECRETARIO,
+  SELEC_DOC,
+  SIGUIENTE,
+  SIN_FILTRO,
+  TIPO_DOCUMENTAL,
+  TRAMITE_OTRO_PROC,
+  TRAMITES,
+  UNSELECT_DOC,
+  URGENTE,
+  VER,
+  VIOLENCIA_GENERO,
+  VISIBILITY,
+  VOLVER,
+  WORD,
+  ZIP,
+  ZOOM_IN,
+};