+++ test/fixture/src/utils/stateNames.js
@@ -1,0 +1,229 @@
+export const ACEPTACION_ASUNTOS = 'aceptacionAsuntos';
+export const ACEPTACION_ESCRITOS = 'aceptacionEscritos';
+export const ACONTECIMIENTOS_PENDIENTES = 'acontecimientosPendientes';
+export const ALTA_MODIFICACION_ACTUACIONES = 'altaModificacionActuaciones';
+export const ACTUACIONES = 'actuaciones';
+export const ADSCRIPCION_ORGANO = 'adscripcionOrgano';
+export const AGENDA_PLAZOS = 'agendaPlazos';
+export const AGENDA_PRESOS = 'agendaPresos';
+export const ALERTA_SIT_INTERVINIENTES = 'alertaSituacionIntervinientes';
+export const ALTA_MODIFICACION_RECURSO_INTERPUESTO = 'altaModificacionRecursoInterpuesto';
+export const ANULACION_ASUNTOS = 'anulacionAsuntos';
+export const ANULACION_REPARTO_ASUNTOS = 'anulacionRepartoAsuntos';
+export const APUNTE_AGENDA_PLAZOS = 'apunteAgendaPlazos';
+export const ASIGNACION_ORIGEN = 'asignacionOrigen';
+export const APERTURA_SECCIONES = 'aperturaSecciones';
+export const ASUNTO_DEVUELTO_A_DECANATO = 'asuntoDevueltoADecanato';
+export const ASUNTO_DEVUELTO_ACEPTADO_RECHAZADO = 'asuntoDevueltoAceptadoRechazado';
+export const ASUNTOS_PENDIENTES = 'asuntosPendientes';
+export const ASUNTOS_REGISTRO = 'asuntosRegistro';
+export const ASUNTOS_STATE = 'asuntoState';
+export const BAJA_ASUNTOS = 'bajaAsuntos';
+export const BAJA_ESCRITOS = 'bajaEscritos';
+export const BLOQUE_DOCUMENTAL = 'bloqueDocumental';
+export const BUSC_ENTIDADES = 'buscEntidades';
+export const BUSC_PERSONAS = 'buscPersonas';
+export const COLECTIVO = 'colectivo';
+export const COLEGIADO = 'colegiado';
+export const CONSULTA_DEVOLUCIONES_ASUNTO = 'consultaDevolucionesAsuntos';
+export const CONSULTA_DEVOLUCIONES_ESCRITOS = 'consultaDevolucionesEscritos';
+export const CONSULTA_DOCS_SAVE_PARAMS_EXPAND_SUBTABLE = 'consultaDocsSaveParamsExpandSubTable';
+export const CONSULTA_EMPRESAS_INSOLVENTES = 'consultaEmpresasInsolventes';
+export const CONSULTA_EQUIPOS_USUARIO = 'consultaEquiposUsuario';
+export const CONSULTA_HCO_ACONTECIMIENTOS = 'consultaHcoAcontecimientos';
+export const CONSULTA_HCO_DEVOLUCIONES = 'consultaHcoDevoluciones';
+export const CONSULTA_MINUTAS = 'consultaMinutas';
+export const CONSULTA_MULTIPLE = 'consultaMultiple';
+export const CONSULTA_NOTIFICACIONES = 'consultaNotificaciones';
+export const CONSULTA_OBJETOS = 'consultaObjetos';
+export const CONSULTA_PE_MED_REQ = 'consultaPeMedReq';
+export const CONSULTA_PERSONACIONES = 'consultaPersonaciones';
+export const CONSULTA_POR_ESCRITOS = 'consultaPorEscritos';
+export const CONSULTA_POR_ORIGEN_CAUSA = 'consultaPorOrigenCausa';
+export const CONSULTA_PROCS_EMPRESAS_INSOLVENTES = 'consultaProcsEmpresasInsolventes';
+export const CONSULTA_PROCS_MATERIA_CLASIFICACION = 'consultaProcsMateriaClasificacion';
+export const CONSULTA_PROCS_POR_PROFESIONAL = 'consultaProcsPorProfesional';
+export const CONSULTA_SMAC = 'consultaSMAC';
+export const COMPOSITOR = 'compositor';
+export const COMPOSITOR_CALLBACK = 'compositorCallback';
+export const COMPOSITOR_COMMON_VALUES = 'compositorCommonValues';
+export const DATA_TABLE = 'dataTable';
+export const DEVOLUCION_ASUNTOS = 'devolucionAsuntos';
+export const DOMICILIO = 'domicilio';
+export const ENTIDAD = 'entidad';
+export const ESCRITO_DEVUELTO_A_DECANATO = 'escritoDevueltoADecanato';
+export const ESCRITO_DEVUELTO_ACEPTADO_RECHAZADO = 'escritoDevueltoAceptadoRechazado';
+export const ESCRITO = 'escrito';
+export const ESCRITOS_PENDIENTES = 'escritosPendientes';
+export const ESCRITOS_SIN_PROC = 'escritosSinProc';
+export const ESCRITOS_TRAMITE = 'escritosTramite';
+export const EXPEDIENTE_ORIGEN = 'expedienteOrigen';
+export const FLUJOS_NAVEGACION = 'flujosNavegacion';
+export const HREF_DELITOS_RELACIONADOS = 'hrefDelitosRelacionados';
+export const INCOACION_ASUNTOS = 'incoacionAsuntos';
+export const INCOACION_PIEZA_SEPARADA = 'incoacionPiezaSeparada';
+export const INTERVINIENTE = 'interviniente';
+export const INTERVINIENTES_SELECTED = 'intervinientesSelected';
+export const ITEMS_TABLA_DESPLEGADOS = 'itemsTablaDesplegados';
+export const LISTADO_PERSONACIONES = 'listadoPersonaciones';
+export const LISTADO_PROCURADORES = 'listadoProcuradores';
+export const MANTENIMIENTO_FASE_ESTADO = 'mantemientoFaseEstado';
+export const MAPA_ASUNTO = 'mapaAsunto';
+export const MEDIDA_CAUTELAR = 'medidaCautelar';
+export const MINUTAS = 'minutas';
+export const MTO_RELACION_DELITO_CLASE_REPARTO = 'mtoRelacionDelitoClaseReparto';
+export const MTO_CLASES_REPARTO = 'mtoClasesReparto';
+export const NUEVO_DOMICILIO = 'nuevoDomicilio';
+export const OPERACION_ESPECIAL_ACUMULACION_DE_PROCEDIMIENTOS =
+  'operacionEspecialAcumulacionDeProcedimientos';
+export const OPERACION_ESPECIAL_APUD_ACTA = 'operacionEspecialApudActa';
+export const OPERACION_ESPECIAL_ARCHIVO_ASUNTO = 'operacionEspecialArchivoAsunto';
+export const OPERACION_ESPECIAL_CIERRE_RECURSO_INTERPUESTO =
+  'operacionEspecialCierreRecursoInterpuesto';
+export const PAGINATION_TABLE_SEARCH_TAREAS = 'paginationTableSearchTareas';
+export const OPERACION_ESPECIAL_TRASPASO_DESGLOSE = 'operacionEspecialTraspasoDesglose';
+export const OPERACION_ESPECIAL_DETALLE_MEDIDAS = 'operacionEspecialDetalleMedidas';
+export const OPERACION_ESPECIAL_DETALLE_PENAS = 'operacionEspecialDetallePenas';
+export const OPERACION_ESPECIAL_DETALLE_REQUERIMIENTOS = 'operacionEspecialDetalleRequerimientos';
+export const OPERACION_ESPECIAL_LIQUIDACIONES_CONDENA = 'operacionEspecialLiquidacionesCondena';
+export const OPERACION_ESPECIAL_LISTADO_LIQUIDACIONES_CONDENA =
+  'operacionEspecialListadoLiquidacionesCondena';
+export const OPERACION_ESPECIAL_MANTENIMIENTO_AGENDA_PLAZOS =
+  'operacionEspecialMantenimientoAgendaPlazos';
+export const OPERACION_ESPECIAL_PENAS_MEDIDAS_REQUERIMIENTOS =
+  'operacionEspecialPenasMedidasRequerimientos';
+export const OPERACION_ESPECIAL_REMISION_RECURSOS_INTERPUESTOS =
+  'operacionEspecialRemisionRecursosInterpuestos';
+export const PENA = 'pena';
+export const PERSONA = 'persona';
+export const PERSONACION = 'personacion';
+export const PROCEDIMIENTO = 'procedimiento';
+export const PROCEDIMIENTOS_INTERVINIENTE = 'procedimientosInterviniente';
+export const REGISTRO_APUD_ACTA = 'registroApudActa';
+export const RELACION_CLASE_REGISTRO_CLASE_REPARTO = 'relacionClaseRegistroClaseReparto';
+export const REQUERIMIENTO = 'requerimiento';
+export const REPRESENTANTE_LEGAL = 'representanteLegal';
+export const SELECCION_PROCEDIMIENTOS = 'seleccionProcedimientos';
+export const SIT_INTERVINIENTE_NO_PENAL = 'situacionIntervNoPenal';
+export const SOLICITUD_MINUTAS = 'solicitudMinutas';
+export const STATE_SIDEBAR = 'stateSideBar';
+export const STORE_URL_PARAMS = 'storeUrlParams';
+export const TAREAS_PENDIENTES_SAVE_STORE = 'tareasPendientes';
+export const TAREAS_PENDIENTES_ITEMS_SELECTED = 'tablePendientesItemSelected';
+export const TAREAS_PENDIENTES_SAVE_PARAMS_EXPAND_SUBTABLE = 'saveParamsExpandSubTable';
+export const TRAMITAR = 'tramitar';
+export const URL_NO_ID = 'urlNoId';
+export const URL_ORIGIN = 'urlOrigin';
+
+export const STATE_NAMES = {
+  ACEPTACION_ASUNTOS,
+  ACEPTACION_ESCRITOS,
+  ACONTECIMIENTOS_PENDIENTES,
+  ALTA_MODIFICACION_ACTUACIONES,
+  ACTUACIONES,
+  ADSCRIPCION_ORGANO,
+  AGENDA_PLAZOS,
+  AGENDA_PRESOS,
+  ALERTA_SIT_INTERVINIENTES,
+  ALTA_MODIFICACION_RECURSO_INTERPUESTO,
+  ANULACION_ASUNTOS,
+  ANULACION_REPARTO_ASUNTOS,
+  APUNTE_AGENDA_PLAZOS,
+  APERTURA_SECCIONES,
+  ASIGNACION_ORIGEN,
+  ASUNTO_DEVUELTO_A_DECANATO,
+  ASUNTO_DEVUELTO_ACEPTADO_RECHAZADO,
+  ASUNTOS_PENDIENTES,
+  ASUNTOS_REGISTRO,
+  ASUNTOS_STATE,
+  BAJA_ASUNTOS,
+  BAJA_ESCRITOS,
+  BLOQUE_DOCUMENTAL,
+  BUSC_ENTIDADES,
+  BUSC_PERSONAS,
+  COLECTIVO,
+  COLEGIADO,
+  CONSULTA_DEVOLUCIONES_ASUNTO,
+  CONSULTA_DEVOLUCIONES_ESCRITOS,
+  CONSULTA_DOCS_SAVE_PARAMS_EXPAND_SUBTABLE,
+  CONSULTA_EMPRESAS_INSOLVENTES,
+  CONSULTA_EQUIPOS_USUARIO,
+  CONSULTA_HCO_ACONTECIMIENTOS,
+  CONSULTA_HCO_DEVOLUCIONES,
+  CONSULTA_MINUTAS,
+  CONSULTA_MULTIPLE,
+  CONSULTA_NOTIFICACIONES,
+  CONSULTA_OBJETOS,
+  CONSULTA_PE_MED_REQ,
+  CONSULTA_PERSONACIONES,
+  CONSULTA_POR_ESCRITOS,
+  CONSULTA_POR_ORIGEN_CAUSA,
+  CONSULTA_PROCS_EMPRESAS_INSOLVENTES,
+  CONSULTA_PROCS_MATERIA_CLASIFICACION,
+  CONSULTA_PROCS_POR_PROFESIONAL,
+  CONSULTA_SMAC,
+  COMPOSITOR,
+  COMPOSITOR_CALLBACK,
+  COMPOSITOR_COMMON_VALUES,
+  DATA_TABLE,
+  DEVOLUCION_ASUNTOS,
+  DOMICILIO,
+  ENTIDAD,
+  ESCRITO_DEVUELTO_A_DECANATO,
+  ESCRITO_DEVUELTO_ACEPTADO_RECHAZADO,
+  ESCRITO,
+  ESCRITOS_PENDIENTES,
+  ESCRITOS_SIN_PROC,
+  ESCRITOS_TRAMITE,
+  EXPEDIENTE_ORIGEN,
+  FLUJOS_NAVEGACION,
+  HREF_DELITOS_RELACIONADOS,
+  INCOACION_ASUNTOS,
+  INCOACION_PIEZA_SEPARADA,
+  INTERVINIENTE,
+  INTERVINIENTES_SELECTED,
+  ITEMS_TABLA_DESPLEGADOS,
+  LISTADO_PERSONACIONES,
+  LISTADO_PROCURADORES,
+  MANTENIMIENTO_FASE_ESTADO,
+  MAPA_ASUNTO,
+  MEDIDA_CAUTELAR,
+  MINUTAS,
+  MTO_RELACION_DELITO_CLASE_REPARTO,
+  MTO_CLASES_REPARTO,
+  NUEVO_DOMICILIO,
+  OPERACION_ESPECIAL_ACUMULACION_DE_PROCEDIMIENTOS,
+  OPERACION_ESPECIAL_APUD_ACTA,
+  OPERACION_ESPECIAL_ARCHIVO_ASUNTO,
+  OPERACION_ESPECIAL_CIERRE_RECURSO_INTERPUESTO,
+  PAGINATION_TABLE_SEARCH_TAREAS,
+  OPERACION_ESPECIAL_TRASPASO_DESGLOSE,
+  OPERACION_ESPECIAL_DETALLE_MEDIDAS,
+  OPERACION_ESPECIAL_DETALLE_PENAS,
+  OPERACION_ESPECIAL_DETALLE_REQUERIMIENTOS,
+  OPERACION_ESPECIAL_LIQUIDACIONES_CONDENA,
+  OPERACION_ESPECIAL_LISTADO_LIQUIDACIONES_CONDENA,
+  OPERACION_ESPECIAL_MANTENIMIENTO_AGENDA_PLAZOS,
+  OPERACION_ESPECIAL_PENAS_MEDIDAS_REQUERIMIENTOS,
+  OPERACION_ESPECIAL_REMISION_RECURSOS_INTERPUESTOS,
+  PENA,
+  PERSONA,
+  PERSONACION,
+  PROCEDIMIENTO,
+  PROCEDIMIENTOS_INTERVINIENTE,
+  REGISTRO_APUD_ACTA,
+  RELACION_CLASE_REGISTRO_CLASE_REPARTO,
+  REQUERIMIENTO,
+  REPRESENTANTE_LEGAL,
+  SELECCION_PROCEDIMIENTOS,
+  SIT_INTERVINIENTE_NO_PENAL,
+  SOLICITUD_MINUTAS,
+  STATE_SIDEBAR,
+  STORE_URL_PARAMS,
+  TAREAS_PENDIENTES_SAVE_STORE,
+  TAREAS_PENDIENTES_ITEMS_SELECTED,
+  TAREAS_PENDIENTES_SAVE_PARAMS_EXPAND_SUBTABLE,
+  TRAMITAR,
+  URL_NO_ID,
+  URL_ORIGIN,
+};