+++ CHANGELOG.md
@@ -9,6 +9,8 @@ All notable changes to the "vue-discovery" extension will be documented in this 
 -   Changes in AST finder dependencies
 -   Now, we can navigate to Cypress Commands definitions
 -   Add improvements to detect and navigate to $refs in a vue page
+-   Add improvements to detect and navigate to constants defined in Vue instance
+    New property 'utilsDirectory' added
 
 ## [0.7.2] 2022-03-31